+++ src/main/java/com/origin/idea/DeepSearchResult.java
@@ -12,13 +12,28 @@ final class DeepSearchResult {
     private final String name;
     private final String path;
     private final String source;
+    private final int lineNumber;
+    private final int columnNumber;
+    private final String preview;
     private final VirtualFile virtualFile;
 
-    DeepSearchResult(SourceType sourceType, String name, String path, String source, VirtualFile virtualFile) {
+    DeepSearchResult(
+            SourceType sourceType,
+            String name,
+            String path,
+            String source,
+            int lineNumber,
+            int columnNumber,
+            String preview,
+            VirtualFile virtualFile
+    ) {
         this.sourceType = sourceType;
         this.name = name;
         this.path = path;
         this.source = source;
+        this.lineNumber = lineNumber;
+        this.columnNumber = columnNumber;
+        this.preview = preview;
         this.virtualFile = virtualFile;
     }
 
@@ -36,6 +51,18 @@ final class DeepSearchResult {
 
     String getSource() {
         return source;
+    }
+
+    int getLineNumber() {
+        return lineNumber;
+    }
+
+    int getColumnNumber() {
+        return columnNumber;
+    }
+
+    String getPreview() {
+        return preview;
     }
 
     VirtualFile getVirtualFile() {